Keep keyboard grab until focus is lost in non-X11
Usually, keyboard grab is released as soon as the mouse pointer leaves
the spice widget. This is done in this way because in X11, no focus
change events are received while the keyboard is grabbed, so it must be
released on widget leave just in case. In other platforms, focus change
events are received, so we can use them to grab/release the keyboard.