FLEXVDI-3536 Fix crash on inactivity timeout
The inactivity timeout is set as soon as its value is received. However,
if the connection is not eventually established and the timeout is
triggered, it tries to access objects that do not exist. This commit:
a) Delays setting the timeout handler until the first window appears.
b) Checks that the active window exists and is a SpiceWindow.