Move desktop and proxy-uri options to General
These two options do not really belong in the Session section.
- proxy-uri is a connection option, like host or port. It should be the
same no matter where the client connects to.
- desktop has a similar degree to user. In fact, in the future, an
interesting feature would be to have sections with a name like
Session:user@host:port/desktop and Devices:user@host:port/desktop,
where options would apply only when connected to that specific host,
with that user and desktop. For this to work, desktop option must not
be in the Session section.