Win-Installer: Get compiler&tools from environment
If CC is defined, use it to get the machine string instead of assuming
its value. This is needed because the last Cerbero toolchain is multilib
and the machine string is always x86_64-w64-mingw32. Then, calculate
tool names after the machine string if they are not defined in the
environment already.