FLEXVDI-3700 Create new windows for additional monitors
Create a new SpiceWindow when additional monitors are enabled. There can
be either up to four monitors in display channel 0 or one monitor (id=0)
in up to four display channels. Besides, when the application goes
fullscreen, the client monitors layout is requested to the guest so they
are activated with the same resolution and position.