Create a PrintJobManager class
Create a class to manage print jobs. Its instance is connected with the
message signal of the flexvdi port. Then, it handles job messages as
usual, but emits a signal when the job is not handled locally. The app
object is in charge of creating and connecting this object, and handling
the "pdf" signal to open the orphan PDF file.
This design will help open the PDF file in the mobile clients.